Move QuestionListPreview to own component
